+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   - -- Duplicating an objective now opens the config for the new objective
   - -- Fixed error from misplaced packager tag for classic
   - -- Loading an objective onto the cursor now clears items or spells on the previously on the cursor
+  - -- Using GameTooltip no longer prevents comparing gear to equipped
+  - -- Fixed an error where changing the tooltip from addon to GameTooltip and back to addon caused blank tooltips
 
 ### 3.0-alpha18